// src/enums/actions.ts
export enum Actions {
  ADD_FOLDER = 'ADD_FOLDER',
  ADD_SUBFOLDER = 'ADD_SUBFOLDER',
  ADD_LINK = 'ADD_LINK',
  DELETE_FOLDER = 'DELETE_FOLDER',
  DELETE_SUBFOLDER = 'DELETE_SUBFOLDER',
  DELETE_LINK = 'DELETE_LINK',
  EDIT_FOLDER = 'EDIT_FOLDER',
  EDIT_SUBFOLDER = 'EDIT_SUBFOLDER',
  EDIT_LINK = 'EDIT_LINK',
}